Added an empty windows Ops Agent plugin to fix tests.
